flask_oauth_jwt_rbac_lab/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Authlib==1.3.1
python-dotenv==1.0.1
pyodbc==5.1.0
requests
